SQLite connection is closed as soon as connection variable becomes not valid in context.
It seems SQLite connection is destroyed as soon as connection variable is no
more necessary any more despite the fact that connection is supplied to OpenID
store. Thus It is not enough to pass connection to store and expect that
it will not be destroyed. Because of that connection object must be acquired
and available through the whole store usage in function. Before that connection
was acquired in _get_consumer function and that was reason why connection was
closed then function was finished.