Bug fix: lower-case authnames consistently.
Authnames were being lower-cased when recovering them from the cookie,
but not when generating them initially.  This resulted — unless the
user’s name was all lower case to start with — in two sessions being
created upon initial login, one of which was ignored thereafter.

This patch adds a new option (lowercase_authname) which determines
wether authnames are either always lower-cased or never lower-cased.
Generally, I think that not lower-casing makes more sense. The default
for this option, however, is to always lower-case the authname in
order to preserve backwards compatibility for those who have
lower-cased user sessions which they would like to preserve.

Also, always uniquify authnames.  When they are lowercased, there’s
always a chance of collision, even when they include the identity URL.