Cookie expiration (and config) fixes. =====================================
Use the [trac]auth_cookie_lifetime config value instead of
[trac]expires to determine cookie lifetime.  Support a setting of zero
(for cookies which last for the browser session).  'Expires' does not
seem to be documented or used by trac 0.12.

In general, avoid declaring options in the [trac] section.  This
has the unwanted side-effects of possibly overriding the default
values and docstrings for these options.

Other buglets fixed
===================

When logging out, do not delete all the users cookies from the
db, just the one for the current browser.

Clean up cookie refreshment when the [openid]timeout option is set.