Clean up openid store schema upgrade handling
There were problems in testing for the existence of db tables by
executing a SELECT on the table and deciding it is not there
if that fails.  Reportedly in some recent trac versions,
environment_needs_upgrade gets a read-only connection which does
not have a rollback() method; and with some types of databases,
the connection (or cursor) won't work after the failed operation
without a rollback.

So now we test for table existence by listing the tables.

And we keep track of the schema state by installing our own row
in the trac 'system' table.  Apparently this is the recommended
best practice.

More discussion on this can be found at:

- http://trac-hacks.org/ticket/9521
- https://groups.google.com/forum/?fromgroups#!topic/trac-dev/RzEqSjCnOhE

Ideas for this implementation from stolen from the TracTags plugin:

- http://trac-hacks.org/browser/tagsplugin/trunk/tractags/db.py?rev=12069